Fix Style/Documentation offenses
- Document undocumented modules and classes
- Remove useless module ExcludeInitialize
- Move Reek::Spec's methods to the logical file
- Put version information in its own module
