+++ lib/reek/spec/should_reek.rb
@@ -21,11 +21,5 @@ module Reek
         "Expected no smells, but got:\n#{rpt}"
       end
     end
-    #
-    # Returns +true+ if and only if the target source code contains smells.
-    #
-    def reek
-      ShouldReek.new
-    end
   end
 end
